tools/requirements.txt
requests>=2.32
Pillow>=12.0
pillow-heif>=0.18
watchdog>=4.0
